Add new non-parallel Docker test suite
