+++ .github/workflows/docker-tests.yml
@@ -1,4 +1,4 @@
-name: Tests
+name: Docker Tests
 on:
   push:
   # Triggers the workflow on labeled PRs only.
@@ -12,12 +12,13 @@ concurrency:
 
 jobs:
   Artifactory-Jenkins-Plugin-Tests:
-    if: (contains(github.event.pull_request.labels.*.name, 'safe to test') || github.event_name == 'push') && github.repository_owner != 'jenkinsci'
+    if: (contains(github.event.pull_request.labels.*.name, 'safe to test') || github.event_name == 'push') && github.repository_owner != 'jenkinsci' && !${JENKINS_DOCKER_TEST_DISABLE}
     name: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-2019]
+        max-parallel: 1
+        os: [ubuntu-latest, macos-latest]
     runs-on: ${{ matrix.os }}
     env:
       GRADLE_OPTS: -Dorg.gradle.daemon=false
@@ -26,11 +27,6 @@ jobs:
         uses: actions/checkout@v3
         with:
           ref: ${{ github.event.pull_request.head.sha }}
-
-      - name: Install Go
-        uses: actions/setup-go@v3
-        with:
-          go-version: 1.20.x
 
       - name: Install npm
         uses: actions/setup-node@v3
@@ -43,37 +39,9 @@ jobs:
           java-version: "8"
           distribution: "temurin"
 
-      - name: Install NuGet
-        uses: nuget/setup-nuget@v1
-        with:
-          nuget-version: 6.x
-
-      - name: Setup Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.x"
-
-      - name: Setup Python virtual environment
-        run: python -m pip install virtualenv && python -m virtualenv venv
-
-      - name: Activate Python virtual env on Windows
-        if: ${{ runner.os == 'Windows' }}
-        run: . venv\Scripts\activate.ps1
-
-      - name: Activate Python virtual env on Posix
-        if: ${{ runner.os == 'macOS' || runner.os == 'Linux' }}
-        run: source venv/bin/activate
-
-      - name: Install Conan
-        run: pip install "conan<2" -q
-
       - name: Configure Linux runner
         if: ${{ runner.os == 'Linux' }}
         run: echo "MAVEN_HOME=$(whereis mvn)" >> $GITHUB_ENV
-
-      - name: Configure Windows runner
-        if: ${{ runner.os == 'Windows' }}
-        run: echo "MAVEN_HOME=$(which.exe mvn)" >> $GITHUB_ENV
 
       - name: Configure MacOS runner
         if: ${{ runner.os == 'macOS' }}
@@ -82,7 +50,7 @@ jobs:
           echo "MAVEN_HOME=$(which mvn)" >> $GITHUB_ENV
           echo "GRADLE_HOME=$(brew info gradle | grep /usr/local/Cellar/gradle | awk '{print $1}')" >> $GITHUB_ENV
 
-      - name: Run tests
+      - name: Run Docker tests
         env:
           JENKINS_PLATFORM_URL: ${{ secrets.PLATFORM_URL }}
           JENKINS_PLATFORM_USERNAME: ${{ secrets.PLATFORM_USER }}
@@ -91,4 +59,4 @@ jobs:
           JENKINS_ARTIFACTORY_DOCKER_PUSH_DOMAIN: ${{ secrets.CONTAINER_REGISTRY }}/docker-local
           JENKINS_ARTIFACTORY_DOCKER_PULL_REPO: docker-remote
           JENKINS_ARTIFACTORY_DOCKER_PUSH_REPO: docker-local
-        run: mvn -V -B -U --no-transfer-progress clean verify -DskipITs=false -Dtest=!**/*DockerITest.java "-Dfindbugs.failOnError=false"
+        run: mvn -V -B -U --no-transfer-progress clean verify -DskipITs=false -Dtest=**/*DockerITest.java "-Dfindbugs.failOnError=false"
